[T-zDT9] Updates notebook to test Jupytext pre-commit hook
In my previous commit, I did not modify the notebook. But, after looking at the documentation, I realised that the `--sync` configuration only executed Jupytext on staged notebooks. Also, the `--sync` option only works when a paired notebook, say in Markdown format, is [already generated](https://jupytext.readthedocs.io/en/latest/using-pre-commit.html) manually using either the command line or as a Python library.

Even after generating a Markdown file, Jupytext complains that `[jupytext] Error: the git index is outdated.` and recommends adding the notebook to Git again. Hence, updated the pre-commit config to not use `--sync` and instead does a simple conversion using `--to` and the notebooks as suggested [here](https://github.com/mwouts/jupytext/issues/442#issuecomment-588393426).

Unfortunately, this also results in a similar issue, and it is quite time-consuming to debug it. So, for now, my best hope for now is to add the notebooks to `.gitignore` and remember to run `jupytext` before committing.

Things 3 task: things:///show?id=AiwvK9rU5EzNkLhrUBzDT9